clients/python/requirements.txt
requests==2.12.0
six==1.10.0
websocket-client==0.37.0
